Add a license
This follows emails from the contributors that they're ok with this
license (or, for the record, possibly BSD 3-Clause).

While the history of this repo indicates that the original work here
began in another repo, from what I can tell of the prior history it
looks like the changes which remotely resemble the code here were
all my own work anyway, so I don't believe there are any other authors
for this work.